@@ -16,7 +16,8 @@
     "@mui/icons-material": "^5.14.9",
     "@mui/material": "^5.14.10",
     "@reduxjs/toolkit": "^1.9.5",
-    "@types/react-dom": "18.2.7",
+    "@types/rc-slider": "^9.3.1",
+    "@types/react-slick": "^0.23.10",
     "@types/styled-components": "^5.1.26",
     "autoprefixer": "10.4.15",
     "axios": "^1.5.0",
@@ -26,8 +27,9 @@
     "eslint-config-next": "13.4.18",
     "framer-motion": "^10.16.4",
     "jwt-decode": "^3.1.2",
-    "next": "^13.4.19",
+    "next": "^13.5.1-canary.1",
     "postcss": "8.4.28",
+    "rc-slider": "^10.2.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.10.1",
@@ -42,6 +44,7 @@
     "slick-carousel": "^1.8.1",
     "socket.io-client": "^4.7.2",
     "styled-components": "^6.0.7",
+    "swiper": "^10.2.0",
     "swr": "^2.2.1",
     "tailwindcss": "3.3.3",
     "typeface-inter": "^3.18.1",
@@ -51,8 +54,7 @@
   "devDependencies": {
     "@types/cookie": "^0.5.1",
     "@types/node": "^20.5.7",
-    "@types/react": "^17.0.65",
-    "@types/react-slick": "^0.23.10",
+    "@types/react-slider": "^1.3.1",
     "@types/redux-thunk": "^2.1.0",
     "babel-plugin-styled-components": "^2.1.4"
   }
